+++ src/server/auditserver/ldcs_audit_server_filemngt.h
@@ -28,7 +28,7 @@ int ldcs_audit_server_filemngt_init (char* location);
 int filemngt_read_file(char *filename, void *buffer, size_t *size, int strip, int *err, int *was_stripped);
 int filemngt_encode_packet(char *filename, void *filecontents, size_t filesize, 
                            int stripped, char **buffer, size_t *buffer_size);
-int filemngt_decode_packet(node_peer_t peer, ldcs_message_t *msg, char *filename, size_t *buffer_size, int *bytes_read, int *is_elf, int *stripped);
+int filemngt_decode_packet(node_peer_t peer, ldcs_message_t *msg, char *filename, size_t *buffer_size, size_t *bytes_read, int *is_elf, int *stripped);
 
 typedef enum {
    clt_unknown,
